Commit with corrections
