+++ src/main/java/org/example/courier/ChecksCourier.java
@@ -1,14 +1,16 @@
-package org.example;
+package org.example.courier;
 
 import io.qameta.allure.Step;
 import io.restassured.response.ValidatableResponse;
 import org.apache.http.HttpStatus;
+import org.example.order.ListOrders;
+import org.example.order.Orders;
 
 import java.util.List;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 
-public class Checks {
+public class ChecksCourier {
     @Step("Create courier")
     public void createCourierSaccess(ValidatableResponse response){
         response.assertThat()
@@ -49,18 +51,5 @@ public class Checks {
                 .statusCode(HttpStatus.SC_BAD_REQUEST)
                 .body("message", equalTo("Недостаточно данных для входа"));
     }
-    @Step("Create order")
-    public Integer checkCreateOrder(ValidatableResponse response){
-        Integer track = response.assertThat()
-                .statusCode(HttpStatus.SC_CREATED)
-                .extract().path("track");
-        return track;
-    }
-    @Step("Check orders not null")
-    public void checkGetOrders(ListOrders listOrders){
-        List<Orders> resultOrders = listOrders.getOrders();
-        boolean result = resultOrders.isEmpty();
-        assert result == false;
 
-    }
 }
