+++ src/main/java/org/example/courier/Creds.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.courier;
 
 public class Creds {
     private String login;
