+++ src/main/java/org/example/order/Order.java
@@ -1,8 +1,7 @@
-package org.example;
+package org.example.order;
 
-public class Orders {
-    private Integer id;
-    private String courierId;
+
+public class Order {
     private String firstName;
     private String lastName;
     private String address;
@@ -10,19 +9,14 @@ public class Orders {
     private String phone;
     private Integer rentTime;
     private String deliveryDate;
-    private String track;
-    private String[] color;
     private String comment;
-    private String createdAt;
-    private String updatedAt;
-    private Integer status;
+    private String[] color;
 
-    public Orders() {
+
+    public Order() {
     }
 
-    public Orders(Integer id, String courierId, String firstName, String lastName, String address, String metroStation, String phone, Integer rentTime, String deliveryDate, String track, String[] color, String comment, String createdAt, String updatedAt, Integer status) {
-        this.id = id;
-        this.courierId = courierId;
+    public Order(String firstName, String lastName, String address, String metroStation, String phone, Integer rentTime, String deliveryDate, String comment, String[] color) {
         this.firstName = firstName;
         this.lastName = lastName;
         this.address = address;
@@ -30,28 +24,8 @@ public class Orders {
         this.phone = phone;
         this.rentTime = rentTime;
         this.deliveryDate = deliveryDate;
-        this.track = track;
-        this.color = color;
         this.comment = comment;
-        this.createdAt = createdAt;
-        this.updatedAt = updatedAt;
-        this.status = status;
-    }
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getCourierId() {
-        return courierId;
-    }
-
-    public void setCourierId(String courierId) {
-        this.courierId = courierId;
+        this.color = color;
     }
 
     public String getFirstName() {
@@ -110,22 +84,6 @@ public class Orders {
         this.deliveryDate = deliveryDate;
     }
 
-    public String getTrack() {
-        return track;
-    }
-
-    public void setTrack(String track) {
-        this.track = track;
-    }
-
-    public String[] getColor() {
-        return color;
-    }
-
-    public void setColor(String[] color) {
-        this.color = color;
-    }
-
     public String getComment() {
         return comment;
     }
@@ -134,27 +92,11 @@ public class Orders {
         this.comment = comment;
     }
 
-    public String getCreatedAt() {
-        return createdAt;
+    public String[] getColor() {
+        return color;
     }
 
-    public void setCreatedAt(String createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    public String getUpdatedAt() {
-        return updatedAt;
-    }
-
-    public void setUpdatedAt(String updatedAt) {
-        this.updatedAt = updatedAt;
-    }
-
-    public Integer getStatus() {
-        return status;
-    }
-
-    public void setStatus(Integer status) {
-        this.status = status;
+    public void setColor(String[] color) {
+        this.color = color;
     }
 }
