+++ src/test/java/TestCreateOrderWithParameterized.java
@@ -1,7 +1,7 @@
 import io.restassured.response.ValidatableResponse;
-import org.example.Checks;
-import org.example.EndPoints;
-import org.example.Order;
+import org.example.order.ApiOrder;
+import org.example.order.ChecksOrder;
+import org.example.order.Order;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
@@ -13,12 +13,12 @@ public class TestCreateOrderWithParameterized {
     public TestCreateOrderWithParameterized(String[] color) {
         this.color = color;
     }
-    private final EndPoints endpoints = new EndPoints();
-    private final Checks check = new Checks();
+    private final ApiOrder endpoint = new ApiOrder();
+    private final ChecksOrder check = new ChecksOrder();
 
     @Parameterized.Parameters
-    public static Object[][][] getTestData(){
-        return new Object[][][] {
+    public static Object[][] getTestData(){
+        return new Object[][] {
                 {new String[]{"BLACK"}},
                 {new String[]{"GREY"}},
                 {new String[]{"BLACK", "GREY"}},
@@ -28,11 +28,10 @@ public class TestCreateOrderWithParameterized {
 
     @Test
     public void createOrder() {
-       // List<String> color = Arrays.asList(arrayColor);
         Order order = new Order("Naruto", "Uchiha", "Konoha, 142 apt.", "4", "+7 800 355 35 35", 5, "2020-06-06", "Saske, come back to Konoha", color);
-        ValidatableResponse orderResponse = endpoints.createOrder(order);
+        ValidatableResponse orderResponse = endpoint.createOrder(order);
         Integer track = check.checkCreateOrder(orderResponse);
         assert track > 0;
-        endpoints.cancelOrder(track);
+        endpoint.cancelOrder(track);
     }
 }
