+++ src/test/java/TestListOrders.java
@@ -1,15 +1,17 @@
 
-import org.example.*;
+import org.example.order.ApiOrder;
+import org.example.order.ChecksOrder;
+import org.example.order.ListOrders;
 import org.junit.Test;
 
 
 public class TestListOrders {
-    private final EndPoints endpoints = new EndPoints();
-    private final Checks check = new Checks();
+    private final ApiOrder endpoint = new ApiOrder();
+    private final ChecksOrder check = new ChecksOrder();
     @Test
     public void listOrdersNotEmpty() {
 
-        ListOrders listOrders = endpoints.getOrders();
+        ListOrders listOrders = endpoint.getOrders();
         check.checkGetOrders(listOrders);
     }
 }
